requirements.txt
aiofiles==0.5.0
alembic==1.5.8
amqp==5.0.6
aniso8601==7.0.0
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
autoflake==1.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==20.8b1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
flake8==3.9.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==2.10
invoke==1.5.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
orjson==3.5.2
pathspec==0.8.1
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pydantic==1.8.1
pyflakes==2.3.1
Pygments==2.8.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
Rx==1.6.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.11
sqlalchemy-mixins==1.3
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
